Refactor virt-v2v
Structure the virt-v2v directory into separate packages for isolation.
