+++ virt-v2v/BUILD.bazel
@@ -1,4 +1,3 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 load(
     "@io_bazel_rules_docker//container:container.bzl",
     "container_image",
@@ -90,34 +89,6 @@ container_image(
     visibility = ["//visibility:public"],
 )
 
-go_library(
-    name = "cold_lib",
-    srcs = [
-        "customize-image.go",
-        "customize-rhel.go",
-        "embed-tool.go",
-        "entrypoint.go",
-        "xml-reader.go",
-    ],
-    embedsrcs = [
-        "scripts/windows/9999-restore_config.ps1",
-        "scripts/windows/9999-restore_config_init.bat",
-        "scripts/windows/firstboot.bat",
-        "scripts/rhel/firstboot/README.md",
-        "scripts/rhel/run/README.md",
-        "scripts/rhel/run/network_config_util.sh",
-        "scripts/rhel/run/network_config_util_test",
-    ],
-    importpath = "github.com/konveyor/forklift-controller/virt-v2v",
-    visibility = ["//visibility:private"],
-)
-
-go_binary(
-    name = "virt-v2v-wrapper",
-    embed = [":cold_lib"],
-    visibility = ["//visibility:public"],
-)
-
 container_image(
     name = "forklift-virt-v2v",
     base = ":virt-v2v-layer",
@@ -129,22 +100,12 @@ container_image(
         "LIBGUESTFS_PATH": "/usr/lib64/guestfs/appliance",
     },
     files = [
-        ":virt-v2v-wrapper",
+        "//cmd:virt-v2v-wrapper",
         "@forklift//cmd/image-converter",
         "@forklift//cmd/virt-v2v-monitor",
     ],
     user = "1001",
     visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "cold_test",
-    srcs = [
-        "cold_test.go",
-        "customize-rhel_test.go",
-    ],
-    data = glob(["testdata/**"]),
-    embed = [":cold_lib"],
 )
 
 rpmtree(
@@ -501,14 +462,4 @@ rpmtree(
         "@zstd-0__1.5.1-2.el9.x86_64//rpm",
     ],
     visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "virt-v2v_test",
-    srcs = [
-        "cold_test.go",
-        "customize-rhel_test.go",
-    ],
-    data = glob(["testdata/**"]),
-    embed = [":virt-v2v_lib"],
 )
