+++ virt-v2v/pkg/customize/rhel_test.go
@@ -1,4 +1,4 @@
-package main
+package customize
 
 import (
 	"fmt"
