+++ virt-v2v/pkg/customize/scripts/rhel/run/network_config_util_test
