+++ virt-v2v/pkg/customize/scripts/windows/9999-restore_config_init.bat
