+++ virt-v2v/pkg/customize/scripts/windows/firstboot.bat
