+++ virt-v2v/pkg/customize/testdata/valid_config.xml
