+++ virt-v2v/pkg/global/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "global",
+    srcs = ["variables.go"],
+    importpath = "github.com/konveyor/forklift-controller/virt-v2v/pkg/global",
+    visibility = ["//visibility:public"],
+)
