+++ virt-v2v/pkg/server/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "server",
+    srcs = ["server.go"],
+    importpath = "github.com/konveyor/forklift-controller/virt-v2v/pkg/server",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//virt-v2v/pkg/global",
+        "//virt-v2v/pkg/utils",
+    ],
+)
