+++ virt-v2v/pkg/utils/BUILD.bazel
@@ -1,0 +1,19 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "utils",
+    srcs = [
+        "command.go",
+        "embed-tool.go",
+        "xml-reader.go",
+    ],
+    importpath = "github.com/konveyor/forklift-controller/virt-v2v/pkg/utils",
+    visibility = ["//visibility:public"],
+    deps = ["//virt-v2v/pkg/global"],
+)
+
+go_test(
+    name = "utils_test",
+    srcs = ["utils_test.go"],
+    embed = [":utils"],
+)
