+++ virt-v2v/pkg/utils/xml-reader.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 import (
 	"encoding/xml"
